feat: review code and tweaks
Let the LLM review its own code and suggest improvements.
